#16 Refactor: 시간 변환 기능 Utils.java로 이동
